fix: subclass ECS logger to handle WSGIRequest
If an exception is raised, Django adds a `request` object to the log
record. ecs_logging doesn't know how to serialize this as JSON so ends
up throwing a logging exception itself.

By subclassing the ECS logger we can add custom handling of this object,
and in our case just stringify it.